test: move ReactNavigateView to test-default
Migrate the last remaining test-react-router source: ReactNavigateView (a
ReactAdapterComponent) and its NavigateView.tsx. test-default gains the
flow-react dependency for the React adapter. Deleted from test-react-router,
which now has no Java/TS sources left.
